complaint-system/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-multipart
aiofiles
pillow
python-magic
pytest
pytest-asyncio
httpx